Use size for host memory size

+++ engine/common/host.h
@@ -15,7 +15,7 @@ typedef struct
 	int argc;
 	char **argv;
 	void *membase;
-	int memsize;
+	size_t memsize;
 } quakeparms_t;
 
 extern quakeparms_t host_parms;
