+++ engine/common/sys_linux.c
@@ -284,7 +284,7 @@ int main (int c, char **v)
 
 	j = COM_CheckParm ("-mem");
 	if (j)
-		parms.memsize = (int)(atof (com_argv[j + 1]) * 1024 * 1024);
+		parms.memsize = (size_t)atoi (com_argv[j + 1]) * 1024 * 1024;
 	parms.membase = malloc (parms.memsize);
 
 	parms.basedir = ".";
